requirements.txt
# Azure SDKs
openai>=1.40.0
azure-search-documents>=11.5.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
httpx>=0.27.0

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# PDF Processing
PyMuPDF>=1.24.0
Pillow>=10.2.0

# Image Processing
opencv-python-headless>=4.9.0
numpy>=1.26.0

# LangChain (optional orchestration)
langchain>=0.2.0
langchain-openai>=0.1.0

# Data
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
